Implement binary search algorithm in c++
